requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
numpy==1.14.5
opencv-python==4.1.0.25
protobuf==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-Levenshtein==0.12.0
scipy==1.3.0
Shapely==1.6.4.post2
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.32.2
Werkzeug==0.15.5
